third_party/WebKit/LayoutTests/inspector/elements/elements-tab-stops.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script type="text/javascript">

var test = function() {

  var maxElements = 1000;
  var elements = 0;
  InspectorTest.domModel.requestDocument(documentRequested);
  function documentRequested() {
    eventSender.keyDown("Tab");
    var startElement = document.deepActiveElement();
    do {
      dumpFocus();
      eventSender.keyDown("Tab");
      elements++;
    } while (startElement !== document.deepActiveElement() && elements < maxElements)

    InspectorTest.addResult('')
    InspectorTest.addResult('Shift+Tab:')
    InspectorTest.addResult('')
    startElement = document.deepActiveElement();
    do {
      dumpFocus();
      eventSender.keyDown("Tab", ['shiftKey']);
      elements++;
    } while (startElement !== document.deepActiveElement() && elements < maxElements)

    if (elements >= maxElements)
      InspectorTest.addResult("FAIL: Unable to complete tab stop cycle.")

    InspectorTest.completeTest();
  }

  function dumpFocus() {
    var element = document.deepActiveElement();
    if (!element) {
      InspectorTest.addResult("null");
      return;
    }
    var name = element.tagName;
    if (element.id)
      name += "#" + element.id;
    if (element.getAttribute('aria-label'))
      name += ":" + element.getAttribute('aria-label')
    else if (element.title)
      name += ":" + element.title
    else if (element.textContent && element.textContent.length < 50) {
      name += ":" + element.textContent.replace('\u200B', '');
    }
    else if (element.className)
      name += "." + element.className.split(" ").join(".");
    InspectorTest.addResult(name);
  }
}
</script>
</head>
<body onload="runTest()">
<p>Tests what elements have focus after pressing tab.</p>
</body>
</html>
